Add shell.nix to provide a dev shell with python3.10 + poetry
Objective: With this (and Nix available on your machine), running
`nix-shell` is sufficient to start developing on FawltyDeps.

The shell hook does the initial setup + install of the poetry
environment, and then enters the poetry shell, leaving you inside the
poetry virtualenv where fawltydeps and its dependencies are already
installed.

Improved-by: Nour El Mawass <nour.elmawass@tweag.io>