Refactor parser sub-package into extract_imports module
This removes a directory level that (for now) distracts more than it
helps. If/when the imports extractor grows to the size where it makes
sense to split it into several modules, then we can reconsider putting
it into its own sub-package.

Rename the parse_imports() function to parse_code() to signify its focus
on the code itself. This makes room for other functions that parse
higher-level structures like files and projects.