extract_imports: Extract parent module when importing parent.child
We're only interested in top-level modules/packages that resolve to
third-party dependencies. As such, when we encounter things like:

    import package.subpackage.module
    from otherpackage.subpackage import module

we will simply return the first word (up to the first period), i.e.
"package" and "otherpackage" in the above example.

We might need something more complicated at a later date.