Introduce a simple/minimal command-line interface
For now this only uses the functionality that has yet been implemented,
i.e. to parse Python code and extract import statements. These is one
command-line option of note: --code, which takes three forms:

  --code <dir>  - Parse Python code found under the given directory
                  (aka. call extract_imports.parse_dir())
  --code <file> - Parse Python code found in the given file
                  (aka. call extract_imports.parse_file())
  --code -      - Parse Python code from standard input
                  (aka. call extract_imports.parse_code())

The intention is to add a future --deps option to tell FawltyDeps where
to find the declared dependencies, as well as other options to control
the comparison and reporting that FawltyDeps will perform.

In addition to --code, there is a pair of -v/--verbose and -q/--quiet
options that regulate the log level: The default log level is WARNING,
and each -v/--verbose increases the log level through INFO to DEBUG.
Likewise, each -q/--quiet decreases the log level through ERROR to
CRITICAL.

Finally, the --help option is automatically supplied by argparse.