Minor revision to black configuration
Black does not offer/allow much configuration, but it makes sense to
set the target-version to the minimum Python version that we want to
support.

Ideally there would also be a way to configure which files we want
formatted (all of them), but black insists on taking this as a command-
line option. Thus we cannot get away with simply running `black` for
formatting, `black .` is the closest we get.