main: Teach command-line interface to list dependencies
This adds the --deps option (similar to --code) to tell fawltydeps where
to look for declared dependencies.

It also introduces two new options to choose which action fawltydeps
should perform:

 --list-imports: List the extracted imports and exit
 --list-deps: List the extracted dependencies and exit

When none of these is given, we currently default to --list-imports.

The available actions will change soon, as will the default action.