main: Flip default action from --list-imports to --check
When none of --check/--list-imports/--list-deps were given, we used to
default to --list-imports. We now instead default to --check, which was
always meant to be the main objective for fawltydeps!

Add a couple of tests to verify that we do the Right Thing when fewer
options are explicitly passed.