pyproject.toml: Make dependency groups self-sufficient
Poetry dependency groups are missing a feature: you cannot specify that
one dependency groups depends on (or includes) another dependency group.

This would be useful e.g. when a module (that is not a hard runtime
dependency) is needed in more than one scenario. For example, in our
project pytest is needed both when running tests (for obvious reasons),
but also when running mypy or pylint (as they need to be able to read
type annotations when following import statements from our test code.)

In these situations we have two alternatives:

  - Include each dependency in exactly one dependency group, and then
    combine dependency groups when installing dependencies for the
    various scenarios, e.g. a developer scenario needs to install
    the nox + test + dev groups, dev is insufficient on its own.

  - Include each dependency in as many dependency groups as needed, so
    that each group is self-sufficient. This makes setting up scenarios
    simpler and less error-prone, but increases duplication inside
    pyproject.toml, and developers must take extra care when updating
    dependencies in this file.

Until now we have gone for the first approach. With this commit, we
switch to the second approach.