CI: Split 'linters' action into 'lint' per Python version and 'format'
It seems that both Mypy and Pylint are sensitive to which Python
interpreter is active while they run, and will provide slightly
different results on different versions. Therefore we want to run both
on top of all our supported Python versions.

Replace the 'linters' action (in GitHub CI, aka. 'lint' in Nox) with two
new actions:

 - lint: Run Mypy and Pylint linters. This action is parametrized with
   our supported Python versions.

 - format: Run isort and black in check-only mode to verify correct
   formatting of our code. This only runs on Python v3.10.

(The 'format' actions now resembles the existing 'reformat' Nox action
that runs isort + black to _fix_ the formatting of the code.)

Add 'lint' and 'format' as separate dependency groups as well, to
complement the CI actions.