main.py: Add helper function to clarify code in perform_actions()
perform_actions() has a lot of considitionals that depend on whether an
action is enabled (i.e. present in the given `actions` set) or not.

These tests were done using a fairly terse syntax:

    if actions & {<one or more actions>}:

which is not very readable unless one is already familiar with the
shorthand for set intersections in Python, plus the fact and
non-empty/empty containers evaluate to true/false in a boolean context.

Aid readability by adding an `is_enabled()` inner helper function, which
does exactly the same, but hopefully in a more readable way.