pyproject.toml: Update pylint config
Update the pylint config directives based on the output from

  pylint --generate-toml-config

It turns out that we have been passing the wrong config directives for
configuring pylint. Amazingly, it seems that pylint has tolerated this,
expect that our [tool.pylint.tests] section has been applied also
outside our tests/ directory.

Note that the buggy section is NOT changed in this commit. It will be
fixed in the next commit, and this commit is merely in preparation for
that.

Also, add a jobs = 4 argument to the pylint config, which makes pylint
analyze modules in parallel. On my work laptop this makes Pylint almost
twice as fast.