Fix pylint misconfiguration by running pylint twice
The [tool.pylint.tests] section in pyproject.toml has not been working
as intended: instead of applying some extra Pylint configuration to the
tests/ subdirectory, it turns out that Pylint does not support this
graduated configuration at all, and the contents of that section was
applied to the entire Pylint run (i.e. ignoring from extra Pyling
directives for all of fawltydeps/ as well).

We could have solved this by using the "pylint-per-file-ignores"
3rd-party add-on package to Pylint, but that package does not work on
Python <v3.8.1.

A much simpler solution is to simply run Pylint _twice_, first on
fawltydeps/ with the stricter configuration, and then finally on tests/
with a few more options disabled.