noxfile.py: Add colored diffs to the output of the format session
When the 'format' action fails in CI (or locally), it simply states that
failed, with no more indication as to exactly what failed, or how to fix
it.

Improve that by telling the `isort` and `black` commands to include
colored diffs showing what changes are needed in order for them to pass.

In order for isort to produce colored diffs, it needs colorama installed
in the 'format' dependency group (and in 'dev', for completeness)