noxfile.py: Run `mypy` without arguments to type-check top-level scripts
We already have appropriate directives in our pyproject.toml to control
Mypy's execution. The command-line args are unnecessary.

However, as a consequence, when we run `mypy` inside the `lint-3.*`
session in Nox, we will now _also_ perform type-checking on any Python
scripts in our top-level directory. This includes the `noxfile.py`
itself, which (since it contains an `import nox` statement) means that
Nox must also be availale in the environment when Mypy is run. Thus we
need to add Nox to the `lint` dependency group.