test_real_projects: Start writing integration tests against real code
The ultimate usefulness of FawltyDeps depends on how well it behaves in
real-world Python projects. One way to verify this is to run FawltyDeps
in known scenarios and check its output.

This commit introduces a test_real_projects module that does just that.
We download stable tarballs of 3rd-party Python projects, unpack them,
run FawltyDeps on them, and verify that it yields the expected output.

For each 3rd-party project, we record the following information:
- A shorthand name, used in the tests
- The URL from which to download a tarball of the project
- The SHA256 checksum of the above tarball
- The set of:
    - imports
    - deps (declared)
    - undeclared deps
    - unused deps
  that we expect to find when analyzing the tarball

For each 3rd-party project, we then run the following test cases:
- Verify that --list-imports returns the expected set of imports
- Verify that --list-deps returns the expected declared deps
- Verify that --check-undeclared returns the expected undeclared deps
- Verify that --check-unused returns the expected unused deps

In this commit, we only add one very simple/trivial 3rd-party project.
More will be added in subsequent commits.

Implementation details: The download/unpack that is done for each
3rd-party project is is potentially costly, therefore:

- All these tests are marked with the new "integration" pytest marker
  (declared in pyproject.toml), and the default pytest run (also
  configured in pyproject.toml) is set to NOT run these tests by
  default. Run pytest with '-m integration' in order run these tests[1]

- We use a relatively new caching mechanism in pytest to cache the
  downloaded and unpacked tarball between test runs. This considerably
  speeds up subsequent test runs.

[1]: To pass pytest args to 'nox', use '--' to separate them from nox
args, e.g. to run the these tests on v3.10 with verbose test output:

    nox -Rs tests-3.10 -- -m integration -v

With-improvements-from: Maria Knorps <maria.knorps@tweag.io>