Run integration tests in GitHub CI
For now, these are NOT run in a separate action, but rather added to the
existing "tests" action.

Take care to place the pytest cache (which now caches 3rd-party projets
between test runs) outside the fawltydeps project itself, and add it to
the GitHub cache, so that every CI run doesn't need to download/unpack
everything from scratch.