Move filtering of stdlib imports from check to extract_imports
In our integration tests we're quickly discovering that --list-imports
returns stdlib imports, which is not in line with our objectives.

Move the filtering of stdlib imports from the check module to the
extract_imports module, so that no stdlib imports are returned from
there. This fixes the output of --list-imports.

Adjust the tests to match:
 - test_compare_imports_to_dependencies.py no longer needs to test
   stdlib imports, as these are no longer returned from extract_imports.
 - test_extract_imports_simple.py needs to add tests for the new
   behavior filtering stdlib imports.
 - test_cmdline.py needs to adjust tests to account for --list-import no
   longer listing stdlib imports.