Convert test_real_projects to read projects/expectations from TOML files
Instead of hardcoding the projects and expectations that we want to test
inside the test_real_project module, we want to store them separately,
preferably in a way that makes it very easy to add new projects/tests.

This commit moves the projects + expectations into TOML files under the
tests/real_projects/ directory.

At test collection time (i.e. when pytest starts) we traverse this dir,
and generate a ThirdPartyProject instance for each .toml file we find.
These instances then dynamically parametrize the four test cases inside
test_real_projects, so that simply by adding a new .toml file, we
automatically gain (up to) four more test cases.

The layout of the .toml is straightforward:

 - A [project] section define the main metadata per project: a name, an
   optional description, the URL for the tarball, and its expected
   SHA256 checksum.

 - Then four more sections define the expected output from running the
   four different FawltyDeps actions on the entire project directory;
   here is a handy correspondence overview:

      TOML section         main.Analysis member    Command-line option
      -----------------    --------------------    -------------------
    - [imports]            .imports                --list-imports
    - [declared_deps]      .declared_deps          --list-deps
    - [undeclared_deps]    .undeclared_deps        --check-undeclared
    - [unused_deps]        .unused_deps            --check-unused

   Each of these four sections are optional. If omitted, the
   corresponding test_*() function in test_real_projects will be skipped
   for this project.

   Finally the contents of each section consist of mapping (quoted) file
   names (relative to the project's root dir) to lists of names. For
   [imports] and [undeclared_deps], these are the Python source files
   and the import names used within. For [declared_deps] and
   [unused_deps], these are the configuration files and the dependency
   names used within.

In the future we might want to expand on this structure to allow several
tests/expectations to be included per project.

Also update the description of these tests in README.md accordingly.