main: Refactor perform_actions() into Analysis class
Instead of collecting both the high-level analysis logic and the
presentation logic inside the perform_action() function, we create an
Analysis class that both allows for better factoring of responsibilities
as well as a more customizable high-level API to our functionality.

Perform a FawltyDeps analysis by calling Analysis.create() with the
required actions, and get back an Analysis object. With this you can
call .print_human_readable() to print the default human-readable report
to the stream of your choice.

This structure allows FawltyDeps to be used in a straightforward and
readable manner from other Python scripts, as exemplifed in
test_real_projects.py:

   from fawltydeps.main import Action, Analysis

   analysis = Analysis.create({Action.LIST_IMPORTS}, code=..., deps=...)

   # Compare expected values _directly_ to the analysis object, instead
   # of instead of having to parse the human-readable output.

It also makes it easy for us to add other output formats in the future,
e.g. Analysis.to_json().

Finally, for --list-imports, we fix the location-based sorting to take
line numbers into account, so that imports are listed in the order they
appear in the files.