test_real_projects: Refactor cached_tarball() into ThirdPartyProject
This is a purely internal refactoring, and should not affect any
behavior.

The cached_tarball() fixture contained a lot of code inside a single
function, and the fact that it also was a pytest fixture made it perhaps
a bit more "magic" than it was worth.

The only reason it was a fixture was because it needed access to the
'request' object (for accessing pytest's cache), and this can just as
well be passed as an argument to a regular function or method.

Split up cached_tarball(), and add it as a separate methods on the
ThirdPartyProject class. Instead of calling cached_tarball() from
within the test case, we can now call project.get_project_dir() instead,
and pass the pytest.Cache object explicitly in. This should hopefully
make it a bit clearer exactly how the caching is performed.