test_real_projects: Ensure cached tarball filename includes project name
Tarballs from tags in Github repos typically do not include the project
name in the filename part of the URL, e.g.

  https://github.com/psf/requests/archive/refs/tags/v2.28.2.tar.gz

has until now been cached as

  .../v2.28.2.tar.gz

In this case

  .../requests-v2.28.2.tar.gz

would clearly be a better filename.

Fix this by prefixing the project name whenever it does not already
occur in the URL's filename.