Ignore dot directories when traversing directory trees
We have noticed that running fawltydeps on itself generates lots of
unexpected output and takes a lot of time. This is because the default
behavior looks for all code and deps under the current dir, which
includes various code and projects that are stored in various "hidden"
directories such as .venv/, .pytest_cache/, .nox/, or similar.
These dirs are not interesting to analyze, and should be ignored.

(In the future we want to make this configurable, but this is a good
default for now.)

Create a walk_dir() helper function in a new fawltydeps.utils module,
and use it from both extract_imports and extract_depenedencies.

Add some tests, too.