types: Replace DependencyComparison with two new classes
Use the new classes UndeclaredDependency and UnusedDependency to store
more of the information that flows through the 'check' module.

This replaces the DependencyComparison class, and instead of returning
instances of that class from check.compare_imports_to_dependencies(),
we now return two lists of UndeclaredDependency and UnusedDependency
objects, respectively.

This will e.g. enable the report of unused dependencies to become more
detailed (in a future commit).