types: Use Location in DeclaredDependency
This replaces the 'location: Path' member of DeclaredDependency with a
`source: Location` member. The new member will contain the same
information as the old (the path of the configuration file where the
associated dependency is declared). In addition the new members enables
us to easily extend declared dependencies with more precise information
(e.g. line number) if and when that information becomes available.