types: Use Location in ParsedImport
This replaces the location, lineno, and cellno members of ParsedImport
with a `source: Location` member. The new member has all the
functionality of the previous three, and simplifies e.g. sorting based
on source location.

One important difference introduced here related to the handling of
sys.stdin as a source for Python code: We used the type Optional[Path],
where None was meant to represent the stdin (or any other non-file)
source, but in practice, the extract_imports module instead used the
ill-formed Path("<stdin>") object to represent stdin. This is now
changed to instead use the literal string "<stdin>", and the type of
Location.path reflects this via the PathOrSpecial type (which is
declared as Union[Path, Literal["<stdin>"]).

Another consequence of this change is that the output of --list-imports
now includes the full location information for the source of the listed
imports, i.e. in addition to the source file, we now also list the cell
number (for notebooks) and the line number.