main: Use relative instead of absolute paths in the default case
When --code or --deps is not specified, they default to the current
directory. Before this commit they did so absolutely, since Path.cwd()
returns the absolute path to the current directory (not simply "./").

The end result is that all the Path objects we generate (when
walking the --code and --deps dirs) and eventually output end up
being absolute paths as well.

Fix this by simply defaulting --code and --deps to Path(".") instead
of Path.cwd(). As a consequence, we now generate only relative paths
internally, including the paths that ultimately end up in the output.

With this, we make no further attempts to "relativize" the arguments
passed to FawltyDeps, rather: if you pass in absolute paths then you
will get absolute paths out, and vice versa. If you pass a mixture of
absolute and relative paths (e.g. `--code=. --deps=$(pwd)`) then you
will get a mixture of absolute and relative paths in the output.