extract_imports: Use isort to filter out more non-external imports
We are only interested in finding imports that refer to _external_ code
(that is, modules/packages for which a corresponding dependency should
be declared).

Until now, we've used isort.place_module() to identify (and ignore)
stdlib modules, but there are more module types we need to ignore:

- from __future__ import ...: This is a special statement handled by the
  Python language itself (used to backport newer language features into
  older language versions).
- from fawltydeps.utils import ...: I.e. self-imports, a module that is
  part of a package will often import other modules from the same
  package.
- from utils import ...: Very similar to the above, a module will often
  import another module found in the same directory (or less commonly,
  found anywhere else in the configured $PYTHONPATH).

We haven't tackled these kinds of imports until now, but as it happens,
we have been sitting on the solution this whole time:

isort.place_module() returns a string that classifies the kind of import
passed to it. AFAICS[1] the possible return values are:

- "FUTURE": from __future import ...
- "STDLIB": Imports of stdlib modules, the one category we have used
            until now for ignoring modules.
- "THIRDPARTY": AFAICS, this corresponds _exactly_ to the external
                imports that we are ultimately interested in.
- "FIRSTPARTY": Corresponds to the last two categories above: self-
                imports, as well as imports of local modules/packages.
- "LOCALFOLDER": AFAICS, this corresponds only to _relative_ imports
                 (from .utils import ...) which we already ignore in
                 our parser (see the latter half of parse_code()).

So by flipping our test from:

    def is_stdlib_import(name: str) -> bool:
        return isort.place_module(name) == "STDLIB"

to:

    def is_external_import(name: str) -> bool:
        return isort.place_module(name) == "THIRDPARTY"

we can - in principle - begin to ignore both __future__ and first-party
imports. However, there are some complications (to be solved in the
following commits):

- For STDLIB imports, isort must be told which _versions_ of the Python
  stdlib should be considered for valid stdlib import names.
- For correctly differentiating between FIRSTPARTY and THIRDPARTY
  imports, isort must be told about the location of project itself, i.e.
  where to look for potential first-party imports.

[1]: https://github.com/PyCQA/isort/blob/5.12.0/isort/sections.py and
https://github.com/PyCQA/isort/blob/5.12.0/isort/place.py contain the
declaration of these strings, as well as the code that implements
isort.place_module(), respectively.