extract_imports: Configure isort to correctly find first-party imports
Recall that isort.place_module() returns one of these strings: "FUTURE"
(for `__future__` imports), "STDLIB" (for stdlib imports, see parent
commit for more details), "THIRDPARTY", "FIRSTPARTY", and "LOCALFOLDER"
(relative imports).

These categories are fairly clearly defined at this point, except for
the exact difference between FIRSTPARTY and THIRDPARTY: In the Python
syntax itself, there is nothing to differentiate these: `import foo`
could refer to a foo.py in the same directory (a FIRSTPARTY import), or
it could refer to a "foo" package from PyPI (a THIRDPARTY import).
Likewise, `from foo_package import bar` could refer to the _current_
package (or a sub-package of the current package), or it could refer to
an external dependency.

It all comes down to the context in which the import is made.
isort.place_module() is able to discern this context, as long as its
`directory` parameter is configured correctly, i.e. to a directory where
FIRSTPARTY imports can be found/resolved.

Anything that isort is unable to resolve/find inside this directory
defaults to being classified as THIRDPARTY.

Since we're specifically looking for THIRDPARTY imports, it is therefore
important that this `directory` is correctly configured, otherwise we
will report first-party imports as undeclared external dependencies!

For now, we fix this by setting `directory` to the path that is passed
to parse_dir(), and otherwise defaulting back to the current directory.

TODO/FIXME: There are some scenarios where this is probably not the
correct solution, for example:

- Passing --code as a single file or a stdin stream will bypass
  parse_dir() altogether, and the current directory will be used to find
  FIRSTPARTY imports. This may or may not be what the user intended.
- The project may do trickery with $PYTHONPATH or other mechanisms to
  alter where it is appropriate to look for FIRSTPARTY imports.

We may have to consider adding command-line (and configuration) options
to FawltyDeps for directly specifying this `directory`, and/or for
including/excluding specific names as FIRSTPARTY imports.