extract_imports: Configure isort to consider ALL versions of the stdlib
In the `requests` project, we find the following code:

    try:  # Python 3
        from http.server import HTTPServer, SimpleHTTPRequestHandler
    except ImportError:  # Python 2
        from BaseHTTPServer import HTTPServer
        from SimpleHTTPServer import SimpleHTTPRequestHandler

where the try-block attempts to import things from the Python 3 stdlib,
and the except clause falls back to attempting the same imports from
their corresponding locations in the Python 2 stdlib.

FawltyDeps, at this point in time, finds _all_ the above import
statements, and has no logic to determine that some of them are dead
code on Python 3. It merely asks isort whether each of them are STDLIB
modules (by passing "http", "BaseHTTPServer", and "SimpleHTTPServer" to
isort.place_module() in sequence).

By default, isort only considers the superset of Python 3 versions when
building its set of stdlib modules, which means that the latter two
modules above are NOT considered part of the stdlib. As a result FD
then reports them as undeclared dependencies.

Fix this by configuring isort to consider _all_ versions (i.e. Python 2
_and_ Python 3) of the Python stdlib.

There is a (small) risk associated with this change: Say that a project
uses an external dependency whose import name overlaps with a stdlib
module from an old version of Python (a version on which this particular
project will never run). In this case FD will mistakenly assume that
this external dependency is an stdlib import:

- If the external dependency is not declared, FD will incorrectly omit
  it from the list of undeclared deps.
- If the external dependency is declared, FD will incorrectly report it
  as an unused dependency.

At this point in time we assume that the number of projects that
contain fallback imports for Python 2 far outweighs the number of
projects that include external dependencies that happen to overlap with
outdated stdlib modules.