extract_declared_dependencies: Parse all kinds of requirements files
Until now we have limited ourselves to parsing requirements-formatted
files called either requirements.txt, requirements-all.txt, or
requirements.in.

With this commit, we will pick up all files whose name contains
"requirements" as a discernible word, and that ends with ".txt" or
".in".

To accomplish this, we must extend the code that selects the appopriate
parser with some simple regex functionality: we still keep a dict that
maps file names to parser functions, but now, if the file name is a
regex pattern, we will match that pattern against the current file
instead of checking for string equality. This is done by the new
get_parser() helper function nested inside
extract_declared_dependencies().

This change results in a few more files being parsed by
extract_declared_dependencies() when run on our real projects:

 - peppy: ./requirements/requirements-doc.txt and
          ./requirements/requirements-test.txt
 - requests: ./requirements-dev.txt