Raise `ArgParseError` in `extract_dependencies` on unsupported file
We previously raised an `ArgParseError` when the code file passed to the main
dispatcher function in `extract_imports.py`, that is, `parse_any_arg`,
was not valid.
On the other hand, the main dispatcher function `extract_dependencies`, in
`extract_dependencies.py` would only log an error message if the passed
dependencies file was in an unsupported format. FawltyDeps would continue
working and would not have an error retrun code.
This change harmonizes the behavior for the two cases, by raising the
`ArgParseError` instead of logging an error and handling it in main with a 2
return code.

- Move `ArgParseError` from `extract_imports.py` to `types.py`
- Raise `ArgParseError` in `extract_dependencies` instead of logging an error
- Change the tests to account for that change in behavior