extract_declared_dependencies: Avoid unwanted normalization of dep names
The .key member of the Requirement object contains a sanitized/normalized
version of the dependency names as it occurs in the requirements.txt file:
- The name is lowercased
- Various "special" characters (incl. "_") has been mapped to "-"

We don't want this normalizattion, as it could create ambiguity between
dependency names (see details in issue #88[1]).

Fortunately, the Requirement object contains a copy (albeit not
documented) of the original dependency name, in the .unsafe_name member.

[1]: https://github.com/tweag/FawltyDeps/issues/88