types: Fix numerical sorting of Location objects
The docstring for Location.__post_init__() states:

    Paths sort alphabetically, the other members sort numerically

However, we did not follow this in practice: e.g. "foo.py:123" would
sort before "foo.py:7" because we used the string representation from
repr() in our sort key, and "123" sorts alphabetically before "7".

Fix this by changing the sort key to a tuple where our numerical members
_remain_ numerical. We map the None cases to the integer -1, which

 (a) is not a valid cell number or line number, and
 (b) sorts before any valid number.

Add a test to verify that the numerical order is obeyed when sorting.