main: Add --json for producing JSON-formatted output
Teach Analysis and the classes in fawltydeps.types to output
JSON-compatible dicts for all the various information that goes into the
Analysis object. Also add Analysis.print_json() to output this in JSON
format to a given file descriptor.

The format of the JSON is intentionally modeled closely on the Analysis
object for now. With time and feedback we may have to reconsider exactly
what information to put into the JSON and how to best structure it.