extract_declared_dependencies.py: Ignore bare URLs in requirements files
According to [1], bare URLs are allowed in requirements.txt files, but
the Requirement.parse() from pkg_resources that we rely on does not
understand them. Furthermore, a URL by itself is not good/interesting
dependency name (at least not as long as we are using the identity
mapping). Hence, for now, simply ignore lines that start with URLs.

[1]: https://pip.pypa.io/en/stable/reference/requirements-file-format/