extract_declared_dependencies: Fix outstanding TODOs in tests
We had some disabled tests that waited for extract_declared_dependencies
to behave in a similar way as extract_imports when encountering missing
or unsupported paths.

Add/enable these tests, and fix extract_declared_dependencies() to make
these tests pass.

The only change in behavior is that when the user pass a --deps argument
that does not exist, we will now raise ArgParseError (which is handled
appropriately in main) instead of raising an OSError when attempting to
walk the non-existent directory.