Hide dataclass members in preparation for JSON refactor
When we start using pydantic to generate JSON, the pydantic JSON encoder
will traverse our classes and include all dataclass fields in the JSON.

We want to make three exceptions, i.e. members to be excluded/elided
from the JSON:
 - Analysis.request
 - Location._sort_key
 - For Location, any other member that is unset, i.e. None

I have tried various techniques to tell pydantic that we want to exclude
these members, but short of refactoring our entire dataclass hierarchy
into a pydantic.BaseModel-based hierarchy (which brings in lots of other
changes, not all good), it seems this is the simplest solution:

Remove the fields-to-be-hidden from the internal .__dataclass_fields__
metadata.

The pydantic.json.pydantic_encoder uses dataclasses.asdict() to traverse
the dataclasses when generating JSON, which in turn uses fields() to
list the available members in a dataclass instance. This in turn relies
on metadata recorded in the .__dataclass_fields__ dict that is stored on
each dataclass _class_. We create a copy of this dict where we _remove_
the to-be-hidden fields, and we then re-attach the copy on the dataclass
_instance_.

This will cause fields() (and hence asdict() and pydantic_encoder()) to
skip past the hidden members when traversing the dataclass instance, and
we achieve the objective of hiding these members from the JSON result.