.github/workflows/lint.yaml: Install our main deps when running Mypy
This was forgotten in commit 50a6185 (PR #137): Although we fixed Nox to
also install the main deps when setting up the 'lint' virtualenv, I
forgot that the GitHub CI does not use the Nox virtualenvs, but sets up
its virtualenv via Poetry instead (because the Poetry virtalenv is
cached, and the Nox ones are not).

This fixes the corresponding setup of Poetry virtualenvs for the lint
target in CI to do the same.