noxfile.py + pyproject.toml: Install our main deps when running Mypy
Mypy (and the other linters) is run on the FawltyDeps code in the
current work tree, and as such there is no need to _install_ FawltyDeps
itself into the virtualenv where the linters are running.

However, Mypy in particular needs to be able to access 3rd-party modules
that are imported by FawltyDeps code in order to verify our
type-correctness.

The result of this, has been that when we add a new main dependency to
FawltyDeps we often also need to add it to the `lint` dependency group,
so that it's also available to Mypy.

This is getting tiresome, especially since a missing module in the
`lint` group manifest as type checking errors that might easily be
misattributed to the 3rd-party module lacking type annotations, and we
as a result end up inserting unnecessary type-ignore or similar
workarounds in our code.

Fix this once and for all by auto-installing FawltyDeps and its
dependencies into the `lint` dependency group. This will make all of
FD's deps always available to Mypy.