check: Add function for finding import names provided by local packages
This adds find_import_names_from_package_name() which will use
importlib.metadata to look up a package name (e.g. a dependency name)
in the current Python environment, and attempt to find the corresponding
import names that this package provides.

If the package does not exist in the local environment, or if
importlib.metadata is not able to find any provided import names, we
return `None`, and expect the caller to fall back to some other
mechanism to map the package name into import name(s).