Include project.name in experiment.name and improve printing
The if-condition that checks if a section should be verified must be
compared against None. Consider:

  if self.imports:

will fail if the imports field is missing from the TOML (as intended),
but it will _also_ fail when the imports field is assigned an empty list
(imports = []). The correct if-condition should be:

  if self.imports is not None:

Same for the three other sections.

Prefix relevant prints with the new experiment name, and also print when
we are _skipping_ the verification of a section in the analysis.