tests: Always call caplog.set_level(...) before using caplog.text
We had a few tests that passed when running the test suite with `nox`,
but failed when running `pytest` directly in the dev shell. The failing
tests were unable to find expected log messages within caplog.text.

It turns out that the log level that caplog is configured with is not
stable/the same in the two different scenarios. Instead of digging
further to get to the bottom of this, we rather make user to call
caplog.set_level() with an appropriate log level in _every_ test that
uses caplog.text.