settings: Introduce Settings class for holding FawltyDeps settings
The new settings module provides a Settings class that builds on top of
the settings management framework provided by pydantic[1].

The Settings class will hold the entirety of FawltyDeps' configuration,
as parsed from:

- command-line options
- fawltydeps_* environment variables
- directives in the [tool.fawltydeps] section in pyproject.toml
- hardcoded defaults that are used when not overridden by the above

By using pydantic's BaseSettings for our Settings, we get automatic
validation/coercion of the members of our Settings object.

For this to work properly, we also need to switch the order of
SpecialPath and Path in our PathOrSpecial union: When pydantic
validates/coerces a union member, it will use the first type in the
union that succeeds. With Path first, Location("<stdin>") would be
automatically coerced to Location(Path("<stdin>")), which is NOT what
we want. Putting SpecialPath (= Literal["<stdin>"]) first in the union
solves this.

[1]: https://docs.pydantic.dev/usage/settings/