Move command-line parser setup into Settings class
The Settings class takes over the management of (most) command-line
arguments. This is in preparation for the Settings module having to
"own" the command-line parsing in order to ensure the proper layering/
cascading of command-line arguments on top of environment variables,
configuration file directives, and the hardcoded defaults in the
Settings class.

We leave the possibility for the `main` module to add its own
command-line arguments for things that do not end up in the Settings
module. For now, this is limited to the -V/--version argument.